main: add get&write_date_time() in loop
